Bug fix: Allow records with only key and no value in ci_cache
There are cases where we need to store in ci_cache
records which have only keys (normaly key:value pairs stored), and
other cases where we do not need to allocate memory for value data.
A such case is the ci_cache used in dnsbl_tables module.
In such cases we should set the ci_cache_entry::val_size to 0 and we must
not call free handlers to free value data.