Bug Fix: viralator mode
We are trying to compute the name of the web object from response headers and 
the requested url after response headers reset and replaced.
This is a bug does not exist in the original Trever patch but added by me 
(Christos) while I polished the code.

This patch resets and replaces the response headers after we had compute the
name of downloaded object.