Bug fix: Restrict permissions on c-icap.ctl Author: Mathieu Parent <sathieu@debian.org> Bug-Debian: http://bugs.debian.org/645122 Patch: https://sourceforge.net/p/c-icap/patches/13/
The named pipe used to send control requests to the c-icap
server has incorrect permissions. That could even be considered
as a security issue as any user on the system can cause some of
the requests to c-icap to not be delivered to it or be
misinterpreted (by reading that file).