Make default interprocessing scheme the "file"
The posix semaphores has the problem that when one process crashed, does not
release/unlock the semaphores for which holds the lock.
This is may cause dead lock to c-icap server, for example in the case a crashed
process holds the locking of accept mutex: The remaining process waiting for
ever the accept mutex to be unlocked.